Return S21 setpoint when the climate control is uninitialized
On a reflash or initial reboot, the climate control by default will
return a target temperature of 0 until the air conditioner is actually
switched on for the first time.

This means that the Home Assistant result looks wrong, and defaults
to supplying a temperature request for the user which isn't valid.

This patch modifies the situation where the climate control has a 0
value, but the code is not going to inspect it's stored values
(because it's off).

This ensures that the AC will always have a sensible setpoint in the
climate control in Home Assistant (namely: whatever the S21 is reporting
which will generally match what the user has on their remote control.)